requirements.txt
# Function dependencies, for example:
# package>=version
pycryptodome==3.14.1
google-cloud-pubsub==2.9.0